Safari-611.1.10/LayoutTests/fast/canvas/canvas-createPattern-video-invalid-expected.txt

Test the handling of invalid arguments in canvas createPattern().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS context.createPattern(undefined, undefined) threw exception TypeError: Type error.
PASS context.createPattern(undefined, null) threw exception TypeError: Type error.
PASS context.createPattern(undefined, '') threw exception TypeError: Type error.
PASS context.createPattern(undefined, 'repeat') threw exception TypeError: Type error.
PASS context.createPattern(null, undefined) threw exception TypeError: Type error.
PASS context.createPattern(null, null) threw exception TypeError: Type error.
PASS context.createPattern(null, '') threw exception TypeError: Type error.
PASS context.createPattern(null, 'repeat') threw exception TypeError: Type error.
PASS successfullyParsed is true

TEST COMPLETE
